requirements.txt
telethon
python-telegram-bot
python-dotenv
requests
flask
websockets
